requirements.txt
asgiref==3.8.1
contourpy==1.2.1
crispy==0.7.4
crispy-bootstrap5==2024.2
cycler==0.12.1
distlib==0.3.8
Django==5.0.6
django-crispy-forms==2.1
django-widget-tweaks==1.5.0
fabio==2024.4.0
filelock==3.13.4
fonttools==4.51.0
h5py==3.11.0
hdf5plugin==4.4.0
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
platformdirs==4.2.0
psycopg2==2.9.9
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
silx==2.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.25.1
